requirements.txt
requests
mss
pyaudio
vosk
pystray
Pillow
pyaudio
pvporcupine
requests
edge-tts
playsound
pystray
Pillow
pyaudio
pvporcupine
requests
edge-tts
playsound
pystray
Pillow
PyQt6>=6.5
pyaudio
pvporcupine
requests
edge-tts
playsound
python-dotenv
python-magic
python-magic-bin
